requirements.txt
gradio>=3.34
python-docx>=0.8.11
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tqdm>=4.65
openai>=1.0.0
numpy>=1.24
pydantic>=1.10
